pyproject.toml
[tool.poetry]
name = "effectivemobile"
version = "0.1.0"
description = ""
authors = ["shedko <v.shedko@kpedev.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
mypy = "^1.11.2"
wemake-python-styleguide = "^0.19.2"
flake8-noqa = "^1.4.0"
python-dotenv = "^1.0.1"
# test
factory-boy = "^3.3.1"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.24.0"
anyio = "^4.6.0"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.3"
twisted = "^24.7.0"
httpx = "^0.27.2"
pytest-anyio = "^0.0.0"


[tool.poetry.extras]
lint = ["wemake-python-styleguide", "flake8-noqa"]
tests = ["pytest", "pytest-asyncio", "pytest-cov", "factory-boy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

